+++ src/components/DateSVG/DateSVG.test.js
@@ -1,13 +1,13 @@
-import React from "react";
-import { shallow } from "enzyme";
-import DateSVG from "./";
+import React from 'react';
+import { shallow } from 'enzyme';
+import DateSVG from './';
 
-describe("DateSVG", () => {
-  it("renders without crashing", () => {
+describe('DateSVG', () => {
+  it('renders without crashing', () => {
     expect(shallow.bind(shallow, <DateSVG />)).not.toThrow();
   });
 
-  it("renders as expected", () => {
+  it('renders as expected', () => {
     expect(shallow(<DateSVG />)).toMatchSnapshot();
   });
 });