+++ src/components/Registration/Registration.test.js
@@ -2,14 +2,12 @@ import React from 'react';
 import { shallow, mount } from 'enzyme';
 import Registration from './';
 import { Provider } from 'react-redux';
-import configureStore from 'redux-mock-store'
-import IdeaCard from "../IdeaCard";
+import configureStore from 'redux-mock-store';
 
 describe('Registration', () => {
-
   const initialState = {};
   const mockStore = configureStore();
-  let store = mockStore(initialState);
+  const store = mockStore(initialState);
 
   it('renders without crashing', () => {
     expect(shallow.bind(shallow, <Provider store={store}><Registration/></Provider>)).not.toThrow();