+++ src/components/SVGSpriteSheet/index.js
@@ -1,14 +1,13 @@
-import React from "react";
-import LocationSymbol from "../LocationSymbol";
-import DateSymbol from "../DateSymbol";
-import AddPersonSymbol from "../AddPersonSymbol";
-
-import "./SVGSpriteSheet.css";
-import LightningSymbol from "../LightningSymbol";
+import React from 'react';
+import LocationSymbol from '../LocationSymbol';
+import DateSymbol from '../DateSymbol';
+import AddPersonSymbol from '../AddPersonSymbol';
+import LightningSymbol from '../LightningSymbol';
+import './SVGSpriteSheet.css';
 
 export const defaults = {
-  className: "SVGSpriteSheet",
-  id: "SVGSpriteSheet",
+  className: 'SVGSpriteSheet',
+  id: 'SVGSpriteSheet'
 };
 
 const SVGSpriteSheet = (props) => (