+++ src/sagas/ideas.test.js
@@ -1,32 +1,26 @@
 import { put, takeLatest } from 'redux-saga/effects';
-import {fetchAllIdeas, catchError, getAllIdeas, ideasActionWatcher} from './ideas'
+import { getAllIdeas, ideasActionWatcher } from './ideas';
+import { fetchAllIdeas } from '../components/App/actions';
 
 describe('ideas saga', () => {
-    test('should dispatch action "FETCH:ALL:IDEAS" ', () => {
-        const generator = ideasActionWatcher();
-        expect(generator.next().value).toEqual(takeLatest('FETCH:ALL:IDEAS', getAllIdeas));
-    });
+  test('should dispatch action "FETCH:ALL:IDEAS" ', () => {
+    const generator = ideasActionWatcher();
+    expect(generator.next().value)
+      .toEqual(takeLatest(fetchAllIdeas, getAllIdeas));
+  });
 
-    test('should dispatch action "RESULT:IDEAS" with result from fetch API', () => {
-        const mockData = [{
-            "team": [{
-                "_id":"5efcb61837631f764b0733b9",
-                "firstName":"hhhjb",
-                "lastName":"cfhdhg",
-                "email":"jndgd@ddd.ddd",
-                "position":"fghgfh",
-                "createdAt":"2020-07-01T16:13:12.084Z",
-                "updatedAt":"2020-07-01T16:13:12.084Z","__v":0
-            }],
-            "_id":"5efca4b11d559156249e7ef1",
-            "name":"ghf",
-            "description":"hgfghfghg",
-            "createdAt":"2020-07-01T14:58:57.947Z",
-            "updatedAt":"2020-07-02T14:44:28.886Z","__v":0
-        }];
-        const generator = getAllIdeas();
-        generator.next();
-        expect(generator.next(mockData).value).toEqual(put({type:"RESULT:IDEAS", value: mockData}));
-    });
-
+  test('should dispatch action "RESULT:IDEAS" with result from fetch API', () => {
+    const mockData = [
+      {
+        'team': [],
+        '_id': '5efca4b11d559156249e7ef1'
+      }];
+    const generator = getAllIdeas();
+    generator.next();
+    expect(generator.next(mockData).value)
+      .toEqual(put({
+        type:'RESULT:IDEAS',
+        value: mockData
+      }));
+  });
 });