+++ src/sagas/index.js
@@ -2,5 +2,5 @@ import { all } from 'redux-saga/effects';
 import { ideasActionWatcher } from './ideas';
 
 export default function* rootSaga() {
-    yield all([ideasActionWatcher()]);
+  yield all([ideasActionWatcher()]);
 }